chore(model-integration): move huggingface embedder to parent directory
