+++ model-integration/src/main/java/ai/vespa/embedding/HuggingFaceEmbedder.java
@@ -1,8 +1,6 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package ai.vespa.embedding.huggingface;
+package ai.vespa.embedding;
 
-import ai.vespa.embedding.PoolingStrategy;
-import ai.vespa.embedding.ModelPathHelper;
 import ai.vespa.modelintegration.evaluator.OnnxEvaluator;
 import ai.vespa.modelintegration.evaluator.OnnxEvaluatorOptions;
 import ai.vespa.modelintegration.evaluator.OnnxRuntime;
